+++ sequence-insights/test/SequenceInsights.Tests/TestData/TestDataGenerator.cs
@@ -1,3 +1,6 @@
+using SequenceInsights.Tests.Models;
+using SequenceInsights.Tests.Utility;
+
 namespace SequenceInsights.Tests.TestData;
 
 public class TestDataGenerator
